# docs/openapi/roadwatch-api.yaml
openapi: 3.0.3
info:
  title: Roadwatch API
  version: 1.0.0
  description: Combined OpenAPI spec for Gateway API (user-facing) and Backend API (auxiliary services). Includes health, auth, complaints, submissions, webhook and analytics endpoints.
servers:
  - url: http://localhost:3100
    description: Local gateway
  - url: http://localhost:4001
    description: Backend auxiliary service
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    NonceResponse:
      type: object
      properties:
        nonce:
          type: string
        issued_at:
          type: string
        expires_at:
          type: string
        ttl_seconds:
          type: integer
    SubmissionResponse:
      type: object
      properties:
        id:
          type: string
        request_id:
          type: string
        verified_status:
          type: string
        message:
          type: string
...
paths:
...
  /submissions/nonce:
    post:
      summary: Generate nonce for image submission
      security:
        - serviceJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                request_id:
                  type: string
                ttl_seconds:
                  type: integer
      responses:
        '200':
          description: Nonce issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NonceResponse'
  /submissions:
    post:
      summary: Submit geotagged image (raw octet-stream)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Submission received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
    get:
      summary: List submissions (authority/admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: request_id
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: list
  /submissions/{id}:
    get:
      summary: Get submission by id (privacy-filtered)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Submission
...